refactor: 공통 컴포넌트의 타입 정의 통합 및 불필요한 파일 제거
