+++ frontend/src/components/@commons/Button/Button.tsx
@@ -1,4 +1,9 @@
-import { ButtonProps } from './Button.types';
+import { SerializedStyles } from '@emotion/react';
+
+export interface ButtonProps extends React.ComponentProps<'button'> {
+  children: React.ReactNode;
+  customCSS?: SerializedStyles | SerializedStyles[];
+}
 
 export default function Button({ children, customCSS, ...props }: ButtonProps) {
   return (
