--- frontend/src/components/@commons/Button/Button.types.ts
@@ -1,6 +1,0 @@
-import { SerializedStyles } from '@emotion/react';
-
-export interface ButtonProps extends React.ComponentProps<'button'> {
-  children: React.ReactNode;
-  customCSS?: SerializedStyles | SerializedStyles[];
-}
