+++ frontend/src/components/@commons/Input/Input.tsx
@@ -1,4 +1,13 @@
-import { InputProps } from './Input.types';
+import { SerializedStyles } from '@emotion/react';
+
+export interface InputProps extends React.ComponentProps<'input'> {
+  value: string;
+  onChange: () => void;
+  placeholder: string;
+  css: SerializedStyles;
+  maxLength: number;
+  minLength: number;
+}
 
 export default function Input({
   value,
