--- frontend/src/components/@commons/Input/Input.types.ts
@@ -1,10 +1,0 @@
-import { SerializedStyles } from '@emotion/react';
-
-export interface InputProps extends React.ComponentProps<'input'> {
-  value: string;
-  onChange: () => void;
-  placeholder: string;
-  css: SerializedStyles;
-  maxLength: number;
-  minLength: number;
-}
