+++ frontend/src/components/@commons/TextArea/TextArea.tsx
@@ -1,4 +1,13 @@
-import { TextAreaProps } from './TextArea.types';
+import { SerializedStyles } from '@emotion/react';
+
+export interface TextAreaProps extends React.ComponentProps<'textarea'> {
+  value: string;
+  onChange: () => void;
+  placeholder: string;
+  css: SerializedStyles;
+  maxLength: number;
+  minLength: number;
+}
 
 export default function TextArea({
   value,
