--- frontend/src/components/@commons/index.txt
@@ -1,1 +1,0 @@
-프로젝트 전체에서 사용되는 공통 컴포넌트(추상화 레벨이 가장 높음)
